# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/site-document:
    get:
      operationId: getPublicSiteDocument
      summary: Resolve an anonymous public URL to an isolated HTML document
      description: |
        Uses only public business/company names. Never loads session ownership,
        account state, licence details or admin metadata. Existing application
        namespaces return an empty 204 so the development server can continue
        its existing routing. Malformed, unknown and ineligible sites share 404.
      parameters:
        - { name: path, in: query, required: true, schema: { type: string, maxLength: 2048 } }
      responses:
        "200":
          description: Shared eligible subscriber public homepage; names and safe generic rendering defaults
          content:
            text/html:
              schema: { type: string }
        "204": { description: Existing application or infrastructure namespace; no public document }
        "404":
          description: Generic unavailable document; no subscriber or eligibility details
          content:
            text/html:
              schema: { type: string }
        "503":
          description: Generic temporary service-unavailable document
          content:
            text/html:
              schema: { type: string }
  /cms/public-website:
    get:
      operationId: getCurrentPublicWebsite
      responses:
        "200":
          description: Authenticated subscriber website configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicWebsiteConfiguration" }
        "401": { description: Authentication required }
        "403": { description: Eligible subscriber required }
    put:
      operationId: updateCurrentPublicWebsite
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicWebsiteUpdate" }
      responses:
        "200":
          description: Confirmed saved configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicWebsiteConfiguration" }
        "400": { description: Invalid input or image ownership }
        "409": { description: Configuration changed since it was loaded }
...
  /cms/public-website/assets:
    post:
      operationId: uploadPublicWebsiteAsset
      requestBody:
        required: true
        content:
          image/png:
            schema: { type: string, format: binary }
          image/jpeg:
            schema: { type: string, format: binary }
      responses:
        "201":
          description: Validated, normalized subscriber-owned image
          content:
            application/json:
              schema:
                type: object
                required: [id, url]
                properties:
                  id: { type: string, format: uuid }
                  url: { type: string }
        "400": { description: Invalid image }
        "413": { description: Image exceeds 5 MB }
...
components:
...
  schemas:
...
    PublicWebsiteUpdate:
      type: object
      additionalProperties: false
      required: [values, revision]
      properties:
        values: { $ref: "#/components/schemas/PublicWebsiteValues" }
        revision: { type: integer, minimum: 0 }
        presentation: { $ref: "#/components/schemas/PublicPresentationValues" }
        presentation_revision: { type: integer, minimum: 0 }
...
    PublicPresentationConfiguration:
      type: object
      required: [values, revision, asset_urls]
      properties:
        values: { $ref: "#/components/schemas/PublicPresentationValues" }
        revision: { type: integer, minimum: 0 }
        asset_urls:
          type: object
          additionalProperties: { type: string }
    PublicPresentationValues:
      type: object
      additionalProperties: false
      required: [logo_strip_enabled, announcements_enabled, custom_html_enabled, custom_html, hero_mode, slider_autoplay, slider_interval, logos, announcements, banners]
      properties:
        logo_strip_enabled: { type: boolean }
        announcements_enabled: { type: boolean }
        logo_strip_settings: { $ref: "#/components/schemas/PublicStripSettings" }
        announcement_ticker_settings: { $ref: "#/components/schemas/PublicTickerSettings" }
        custom_html_enabled: { type: boolean }
        custom_html: { type: string, maxLength: 4096 }
        hero_mode: { type: string, enum: [classic, banner] }
        slider_autoplay: { type: boolean }
        slider_interval: { type: integer, enum: [3, 5, 7, 10] }
        logos:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/PublicPartnerLogo" }
        announcements:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/PublicAnnouncement" }
        banners:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/PublicBanner" }
    PublicStripSettings:
      type: object
      additionalProperties: false
      required: [display, speed, direction, pause_on_hover]
      properties:
        display: { type: string, enum: [static, moving] }
        speed: { type: string, enum: [slow, normal, fast] }
        direction: { type: string, enum: [left, right] }
        pause_on_hover: { type: boolean }
    PublicTickerSettings:
      type: object
      additionalProperties: false
      required: [display, speed, direction, pause_on_hover, background_color, text_color, separator]
      properties:
        display: { type: string, enum: [static, moving] }
        speed: { type: string, enum: [slow, normal, fast] }
        direction: { type: string, enum: [left, right] }
        pause_on_hover: { type: boolean }
        background_color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        text_color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        separator: { type: string, maxLength: 16 }
    PublicPartnerLogo:
      type: object
      additionalProperties: false
      required: [id, asset_id, label, destination, new_tab, enabled]
      properties:
        id: { type: string, format: uuid }
        asset_id: { type: string, format: uuid }
        label: { type: string, maxLength: 120 }
        destination: { type: string, maxLength: 512 }
        new_tab: { type: boolean }
        enabled: { type: boolean }
    PublicAnnouncement:
      type: object
      additionalProperties: false
      required: [id, enabled, text, destination, background_color, text_color, movement, direction, speed]
      properties:
        id: { type: string, format: uuid }
        enabled: { type: boolean }
        text: { type: string, minLength: 1, maxLength: 500 }
        icon_text: { type: string, maxLength: 32 }
        destination: { type: string, maxLength: 512 }
        background_color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        text_color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        movement: { type: string, enum: [static, scrolling] }
        direction: { type: string, enum: [left, right] }
        speed: { type: string, enum: [slow, normal, fast] }
    PublicBanner:
      type: object
      additionalProperties: false
      required: [id, asset_id, alt_text, destination, enabled]
      properties:
        id: { type: string, format: uuid }
        asset_id: { type: string, format: uuid }
        alt_text: { type: string, maxLength: 180 }
        destination: { type: string, maxLength: 512 }
        enabled: { type: boolean }
    PublicWebsiteValues:
      type: object
      additionalProperties: false
      required: [display_name, business_description, primary_color, hero_badge, hero_title, hero_description, logo_asset_id, hero_asset_id, primary_cta_label, primary_cta_destination, secondary_cta_label, secondary_cta_destination]
      properties:
        display_name: { type: string, maxLength: 120 }
        business_description: { type: string, maxLength: 500 }
        primary_color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        hero_badge: { type: string, maxLength: 120 }
        hero_title: { type: string, maxLength: 180 }
        hero_description: { type: string, maxLength: 1000 }
        logo_asset_id: { type: [string, "null"], format: uuid }
        hero_asset_id: { type: [string, "null"], format: uuid }
        primary_cta_label: { type: string, maxLength: 60 }
        primary_cta_destination: { type: string, maxLength: 512 }
        secondary_cta_label: { type: string, maxLength: 60 }
        secondary_cta_destination: { type: string, maxLength: 512 }
    PublicWebsiteConfiguration:
      type: object
      required: [values, defaults, revision, public_slug, public_url, preview_url, logo_url, hero_image_url]
      properties:
        values: { $ref: "#/components/schemas/PublicWebsiteValues" }
        presentation: { $ref: "#/components/schemas/PublicPresentationConfiguration" }
        defaults: { $ref: "#/components/schemas/PublicWebsiteValues" }
        revision: { type: integer, minimum: 0 }
        public_slug: { type: string }
        public_url: { type: string }
        preview_url: { type: string }
        logo_url: { type: [string, "null"] }
        hero_image_url: { type: [string, "null"] }
...